pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asana-mcp"
version = "0.1.0"
description = "Asana MCP server exposing the Asana REST API as structured tools."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Asana MCP", email = "noreply@example.com" }
]
license = { text = "MIT" }
dependencies = [
  "fastmcp>=0.1.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "aiohttp>=3.9.0",
]

[project.scripts]
asana-mcp = "src.main:main"

